Removing unused appName
Setting appName("Adult Census Income") on the Spark Context may not work here as expected. The spark context is not started here, and so any settings on the sparkcontext will not take effect. Adding this setting here can be confusing and lead to the user trying to set other settings here. All settings must be done in the spark_dependencies.yml